+++ languages/__init__.py
@@ -6,4 +6,20 @@ from . import en
 from . import es
 
 ### EDITABLE VARIABLES #########################################################
-default = en.EN
+default_lang = en.EN
+
+### FUNCTIONS ##################################################################
+def set_lang(str):
+    lang = default_lang
+    if str == "en.EN":
+        print("[INFO] "+lang.set_lang_info+" '"+str+"'")
+        lang = en.EN
+    elif str == "es.ES":
+        print("[INFO] "+lang.set_lang_info+" '"+str+"'")
+        lang = es.ES
+    elif str == "es.GL":
+        print("[INFO] "+lang.set_lang_info+" '"+str+"'")
+        lang = es.GL
+    else:
+        print("[FAIL] '"+str+"' "+lang.set_lang_error)
+    return lang
